requirements.txt
Flask_OAuthlib==0.9.6
Flask==2.0.1
python_twitch_client==0.7.1
mysql
bcrypt
twitch-python
gevent
thefuzz[speedup]